@@ -3,22 +3,25 @@
   "version": "0.1.0",
   "private": true,
   "dependencies": {
-    "@material-ui/core": "^4.11.2",
-    "@material-ui/icons": "^4.11.2",
+    "@emotion/react": "^11.7.1",
+    "@emotion/styled": "^11.6.0",
+    "@mui/icons-material": "^5.2.5",
+    "@mui/material": "^5.2.5",
+    "@mui/styles": "^5.2.3",
     "accounting": "^0.4.1",
-    "axios": "^0.19.0",
-    "dateformat": "^3.0.3",
+    "axios": "^0.24.0",
+    "dateformat": "^5.0.2",
     "local-storage": "^2.0.0",
-    "moment": "^2.24.0",
-    "react": "^16.12.0",
-    "react-datepicker": "^2.10.1",
-    "react-dom": "^16.12.0",
-    "react-dropdown": "^1.6.4",
-    "react-moment": "^0.9.7",
-    "react-router-dom": "^5.1.2",
-    "react-scripts": "3.2.0",
-    "react-select": "^3.0.8",
-    "react-select-search": "^0.10.1"
+    "moment": "^2.29.1",
+    "react": "^17.0.2",
+    "react-datepicker": "^4.5.0",
+    "react-dom": "^17.0.2",
+    "react-dropdown": "^1.9.2",
+    "react-moment": "^1.1.1",
+    "react-router-dom": "^6.2.1",
+    "react-scripts": "5.0.0",
+    "react-select": "^5.2.1",
+    "react-select-search": "^3.0.9"
   },
   "scripts": {
     "start": "react-scripts start",
