@@ -3,11 +3,11 @@ import accounting from 'accounting';
 import '../../css/Results.css';
 import TimeBar from './TimeBar.js';
 import moment from 'moment';
-import Button from '@material-ui/core/Button';
-import Grid from '@material-ui/core/Grid';
-import Box from '@material-ui/core/Box';
-import { withStyles } from '@material-ui/core/styles';
-import CircularProgress from '@material-ui/core/CircularProgress';
+import Button from '@mui/material/Button';
+import Grid from '@mui/material/Grid';
+import Box from '@mui/material/Box';
+import { withStyles } from '@mui/styles';
+import CircularProgress from '@mui/material/CircularProgress';
 import face_screaming from '../../imgs/fomo/face_screaming.png';
 import loudly_crying from '../../imgs/fomo/loudly_crying.png';
 import fear_full from '../../imgs/fomo/fear_full.png';
